requirements.txt
requests~=2.24.0
grafana_api
ldap3~=2.7
mockito~=1.2.1
mock~=4.0.2
PyYAML~=5.3.1
setuptools~=49.2.0